+++ sample/app/src/main/AndroidManifest.xml
@@ -10,14 +10,14 @@
     android:theme="@style/AppTheme"
     tools:ignore="GoogleAppIndexingWarning">
 
-    <activity android:name=".MyActivity">
+    <activity android:name="com.novoda.staticanalysisplugin.sample.MyActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
         <category android:name="android.intent.category.LAUNCHER" />
       </intent-filter>
     </activity>
 
-    <activity android:name=".SomeOtherActivity" />
+    <activity android:name="com.novoda.staticanalysisplugin.sample.SomeOtherActivity" />
 
   </application>
