--- sample/app/src/main/java/com/novoda/buildpropertiesplugin/sample/MyClass.java
@@ -1,4 +1,0 @@
-package com.novoda.buildpropertiesplugin.sample;
-
-public class MyClass {
-}
