+++ sample/app/src/main/java/com/novoda/staticanalysisplugin/sample/MyClass.java
@@ -1,0 +1,4 @@
+package com.novoda.staticanalysisplugin.sample;
+
+public class MyClass {
+}
