+++ sample/app/src/main/java/com/novoda/staticanalysisplugin/sample/SomeOtherActivity.java
@@ -1,10 +1,12 @@
-package com.novoda.buildpropertiesplugin.sample;
+package com.novoda.staticanalysisplugin.sample;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.Toast;
+
+import com.novoda.buildpropertiesplugin.sample.R;
 
 public class SomeOtherActivity extends AppCompatActivity {
